+++ simplebluez/include/simplebluez/interfaces/GattService1.h
@@ -3,23 +3,22 @@
 #include <simpledbus/advanced/Interface.h>
 #include <simpledbus/advanced/InterfaceRegistry.h>
 #include <string>
+#include "simpledbus/base/HolderUtils.h"
 
 namespace SimpleBluez {
 
 class GattService1 : public SimpleDBus::Interface {
   public:
     GattService1(std::shared_ptr<SimpleDBus::Connection> conn, std::shared_ptr<SimpleDBus::Proxy> proxy);
-    virtual ~GattService1() = default;
+    virtual ~GattService1();
 
     // ----- METHODS -----
 
     // ----- PROPERTIES -----
-    std::string UUID();
-
-  protected:
-    void property_changed(std::string option_name) override;
-
-    std::string _uuid;
+    Property<std::string>& UUID = property<std::string>("UUID");
+    Property<bool>& Primary = property<bool>("Primary");
+    CustomProperty<std::vector<std::string>>& Characteristics = property<std::vector<std::string>>(
+        "Characteristics", SimpleDBus::HolderUtils::from_string_array, SimpleDBus::HolderUtils::to_string_array);
 
   private:
     static const SimpleDBus::AutoRegisterInterface<GattService1> registry;