+++ simpledbus/include/simpledbus/interfaces/ObjectManager.h
@@ -3,19 +3,20 @@
 #include <simpledbus/advanced/Interface.h>
 #include <simpledbus/advanced/InterfaceRegistry.h>
 
-#include <functional>
+#include <kvn/kvn_safe_callback.hpp>
 
 namespace SimpleDBus::Interfaces {
 
 class ObjectManager : public Interface {
   public:
     ObjectManager(std::shared_ptr<Connection> conn, std::shared_ptr<Proxy> proxy);
-    virtual ~ObjectManager() = default;
+    virtual ~ObjectManager();
 
-    // Names are made matching the ones from the DBus specification
-    Holder GetManagedObjects(bool use_callbacks = false);
-    std::function<void(std::string path, Holder options)> InterfacesAdded;
-    std::function<void(std::string path, Holder options)> InterfacesRemoved;
+    Holder GetManagedObjects();
+
+    // ----- SIGNALS -----
+    kvn::safe_callback<void(std::string path, Holder options)> InterfacesAdded;
+    kvn::safe_callback<void(std::string path, Holder options)> InterfacesRemoved;
 
     void message_handle(Message& msg) override;