+++ simpledbus/src/interfaces/Properties.cpp
@@ -4,10 +4,10 @@
 using namespace SimpleDBus;
 using namespace SimpleDBus::Interfaces;
 
-const AutoRegisterInterface<Properties> Properties::registry {
+const AutoRegisterInterface<Properties> Properties::registry{
     "org.freedesktop.DBus.Properties",
     // clang-format off
-    [](std::shared_ptr<Connection> conn, std::shared_ptr<Proxy> proxy) -> std::shared_ptr<SimpleDBus::Interface> {
+    [](std::shared_ptr<Connection> conn, std::shared_ptr<Proxy> proxy) -> std::shared_ptr<Interface> {
         return std::make_shared<Properties>(conn, proxy);
     }
     // clang-format on
@@ -15,6 +15,11 @@ const AutoRegisterInterface<Properties> Properties::registry {
 
 Properties::Properties(std::shared_ptr<Connection> conn, std::shared_ptr<Proxy> proxy)
     : Interface(conn, proxy, "org.freedesktop.DBus.Properties") {}
+
+// IMPORTANT: The destructor is defined here (instead of inline) to anchor the vtable to this object file.
+// This prevents the linker from stripping this translation unit and ensures the static 'registry' variable is
+// initialized at startup.
+Properties::~Properties() = default;
 
 Holder Properties::Get(const std::string& interface_name, const std::string& property_name) {
     Message query_msg = Message::create_method_call(_bus_name, _path, "org.freedesktop.DBus.Properties", "Get");
@@ -95,15 +100,9 @@ void Properties::message_handle(Message& msg) {
         std::string iface_name = interface_h.get_string();
 
         std::shared_ptr<Interface> interface = proxy()->interface_get(iface_name);
-        SimpleDBus::Holder properties = SimpleDBus::Holder::create_dict();
-        {
-            std::scoped_lock lock(interface->_property_update_mutex);
-            for (const auto& [key, value] : interface->_properties) {
-                properties.dict_append(SimpleDBus::Holder::Type::STRING, key, value);
-            }
-        }
+        Holder properties = interface->handle_property_get_all();
 
-        SimpleDBus::Message reply = SimpleDBus::Message::create_method_return(msg);
+        Message reply = Message::create_method_return(msg);
         reply.append_argument(properties, "a{sv}");
         _conn->send(reply);
 
@@ -117,25 +116,18 @@ void Properties::message_handle(Message& msg) {
 
         std::shared_ptr<Interface> interface = proxy()->interface_get(iface_name);
 
-        bool property_exists = false;
-        Holder property_value;
-        {
-            std::scoped_lock lock(interface->_property_update_mutex);
-            property_exists = interface->_properties.find(property_name) != interface->_properties.end();
-            if (property_exists) {
-                property_value = interface->_properties[property_name];
-            }
-        }
+        bool property_exists = interface->property_exists(property_name);
 
-        if (!property_exists) {
-            SimpleDBus::Message reply = SimpleDBus::Message::create_error(msg, "org.freedesktop.DBus.Error.InvalidArgs", "Property not found");
+        if (property_exists) {
+            Holder property_value = interface->handle_property_get(property_name);
+            Message reply = Message::create_method_return(msg);
+            reply.append_argument(property_value, "v");
             _conn->send(reply);
-            return;
+        } else {
+            Message reply = Message::create_error(msg, "org.freedesktop.DBus.Error.InvalidArgs",
+                                                                          "Property not found");
+            _conn->send(reply);
         }
-
-        SimpleDBus::Message reply = SimpleDBus::Message::create_method_return(msg);
-        reply.append_argument(property_value, "v");
-        _conn->send(reply);
 
     } else if (msg.is_method_call(_interface_name, "Set")) {
         Holder interface_h = msg.extract();
@@ -149,16 +141,9 @@ void Properties::message_handle(Message& msg) {
         Holder value_h = msg.extract();
 
         std::shared_ptr<Interface> interface = proxy()->interface_get(iface_name);
-        {
-            std::scoped_lock lock(interface->_property_update_mutex);
-            // Only update the property if it exists.
-            // TODO: Should we send an error message if the property doesn't exist?
-            if (interface->_properties.find(property_name) != interface->_properties.end()) {
-                interface->_properties[property_name] = value_h;
-            }
-        }
+        interface->handle_property_set(property_name, value_h);
 
-        SimpleDBus::Message reply = SimpleDBus::Message::create_method_return(msg);
+        Message reply = Message::create_method_return(msg);
         _conn->send(reply);
 
     } else if (msg.is_signal(_interface_name, "PropertiesChanged")) {
@@ -174,6 +159,6 @@ void Properties::message_handle(Message& msg) {
             return;
         }
 
-        proxy()->interface_get(iface_name)->signal_property_changed(changed_properties, invalidated_properties);
+        proxy()->interface_get(iface_name)->handle_properties_changed(changed_properties, invalidated_properties);
     }
 }